+++ dune/istl/ildl.hh
@@ -1,6 +1,8 @@
 #ifndef DUNE_ISTL_ILDL_HH
 #define DUNE_ISTL_ILDL_HH
 
+#include <dune/common/scalarvectorview.hh>
+#include <dune/common/scalarmatrixview.hh>
 #include "ilu.hh"
 
 /**
@@ -29,8 +31,16 @@ namespace Dune
     }
   }
 
+  template< class K >
+  inline static void bildl_subtractBCT ( const K &B, const K &CT, K &A,
+                                         typename std::enable_if_t<Dune::IsNumber<K>::value>* sfinae = nullptr )
+  {
+    A -= B * CT;
+  }
+
   template< class Matrix >
-  inline static void bildl_subtractBCT ( const Matrix &B, const Matrix &CT, Matrix &A )
+  inline static void bildl_subtractBCT ( const Matrix &B, const Matrix &CT, Matrix &A,
+                                         typename std::enable_if_t<!Dune::IsNumber<Matrix>::value>* sfinae = nullptr )
   {
     for( auto i = A.begin(), iend = A.end(); i != iend; ++i )
     {
@@ -114,12 +124,12 @@ namespace Dune
         const auto &A_k = A[ ik.index() ];
 
         auto B = A_ik;
-        A_ik.rightmultiply( *A_k.find( ik.index() ) );
+        Impl::asMatrix(A_ik).rightmultiply( Impl::asMatrix(*A_k.find( ik.index() )) );
         bildl_subtractBCT( B, A_ik, A_ii );
       }
       try
       {
-        A_ii.invert();
+        Impl::asMatrix(A_ii).invert();
       }
       catch( const Dune::FMatrixError &e )
       {
@@ -142,7 +152,10 @@ namespace Dune
       const auto &A_i = *i;
       v[ i.index() ] = d[ i.index() ];
       for( auto ij = A_i.begin(); ij.index() < i.index(); ++ij )
-        (*ij).mmv( v[ ij.index() ], v[ i.index() ] );
+      {
+        auto&& vi = Impl::asVector( v[ i.index() ] );
+        Impl::asMatrix(*ij).mmv(Impl::asVector( v[ ij.index() ] ), vi);
+      }
     }
 
     // solve D w = v, note: diagonal stores Dii^{-1}
@@ -155,8 +168,16 @@ namespace Dune
         const auto &A_i = *i;
         const auto ii = A_i.beforeEnd();
         assert( ii.index() == i.index() );
-        auto rhs = v[ i.index() ];
-        ii->mv( rhs, v[ i.index() ] );
+        // We need to be careful here: Directly using
+        // auto rhs = Impl::asVector(v[ i.index() ]);
+        // is not OK in case this is a proxy. Hence
+        // we first have to copy the value. Notice that
+        // this is still not OK, if the vector type itself returns
+        // proxy references.
+        auto rhsValue = v[ i.index() ];
+        auto&& rhs = Impl::asVector(rhsValue);
+        auto&& vi = Impl::asVector( v[ i.index() ] );
+        Impl::asMatrix(*ii).mv(rhs, vi);
       }
     }
     else
@@ -168,8 +189,16 @@ namespace Dune
         const auto &A_i = *i;
         const auto ii = A_i.find( i.index() );
         assert( ii.index() == i.index() );
-        auto rhs = v[ i.index() ];
-        ii->mv( rhs, v[ i.index() ] );
+        // We need to be careful here: Directly using
+        // auto rhs = Impl::asVector(v[ i.index() ]);
+        // is not OK in case this is a proxy. Hence
+        // we first have to copy the value. Notice that
+        // this is still not OK, if the vector type itself returns
+        // proxy references.
+        auto rhsValue = v[ i.index() ];
+        auto&& rhs = Impl::asVector(rhsValue);
+        auto&& vi = Impl::asVector( v[ i.index() ] );
+        Impl::asMatrix(*ii).mv(rhs, vi);
       }
     }
 
@@ -179,7 +208,10 @@ namespace Dune
     {
       const auto &A_i = *i;
       for( auto ij = A_i.begin(); ij.index() < i.index(); ++ij )
-        (*ij).mmtv( v[ i.index() ], v[ ij.index() ] );
+      {
+        auto&& vij = Impl::asVector( v[ ij.index() ] );
+        Impl::asMatrix(*ij).mmtv(Impl::asVector( v[ i.index() ] ), vij);
+      }
     }
   }