+++ dune/istl/novlpschwarz.hh
@@ -89,6 +89,10 @@ namespace Dune {
      * data points. (E.~g. OwnerOverlapCommunication )
      */
     NonoverlappingSchwarzOperator (const matrix_type& A, const communication_type& com)
+      : _A_(stackobject_to_shared_ptr(A)), communication(com), buildcomm(true)
+    {}
+
+    NonoverlappingSchwarzOperator (std::shared_ptr<const matrix_type> A, const communication_type& com)
       : _A_(A), communication(com), buildcomm(true)
     {}
 
@@ -115,7 +119,7 @@ namespace Dune {
     //! get matrix via *
     virtual const matrix_type& getmat () const
     {
-      return _A_;
+      return *_A_;
     }
 
     void novlp_op_apply (const X& x, Y& y, field_type alpha) const
@@ -149,7 +153,7 @@ namespace Dune {
 
         // for each local index make multimap rimap:
         // key: local index i, data: pair of process that knows i and pointer to RI entry
-        for (RowIterator i = _A_.begin(); i != _A_.end(); ++i)
+        for (RowIterator i = _A_->begin(); i != _A_->end(); ++i)
           if (mask[i.index()] == 0)
             for (RIIterator remote = ri.begin(); remote != ri.end(); ++remote) {
               RIL& ril = *(remote->second.first);
@@ -165,12 +169,12 @@ namespace Dune {
             }
 
         int iowner = 0;
-        for (RowIterator i = _A_.begin(); i != _A_.end(); ++i) {
+        for (RowIterator i = _A_->begin(); i != _A_->end(); ++i) {
           if (mask[i.index()] == 0) {
             std::map<int,int>::iterator it = owner.find(i.index());
             iowner = it->second;
             std::pair<RIMapit, RIMapit> foundiit = rimap.equal_range(i.index());
-            for (ColIterator j = _A_[i.index()].begin(); j != _A_[i.index()].end(); ++j) {
+            for (ColIterator j = (*_A_)[i.index()].begin(); j != (*_A_)[i.index()].end(); ++j) {
               if (mask[j.index()] == 0) {
                 bool flag = true;
                 for (RIMapit foundi = foundiit.first; foundi != foundiit.second; ++foundi) {
@@ -203,10 +207,10 @@ namespace Dune {
       }
 
       //compute alpha*A*x nonoverlapping case
-      for (RowIterator i = _A_.begin(); i != _A_.end(); ++i) {
+      for (RowIterator i = _A_->begin(); i != _A_->end(); ++i) {
         if (mask[i.index()] == 0) {
           //dof doesn't belong to process but is border (not ghost)
-          for (ColIterator j = _A_[i.index()].begin(); j != _A_[i.index()].end(); ++j) {
+          for (ColIterator j = (*_A_)[i.index()].begin(); j != (*_A_)[i.index()].end(); ++j) {
             if (mask[j.index()] == 1) //j is owner => then sum entries
               (*j).usmv(alpha,x[j.index()],y[i.index()]);
             else if (mask[j.index()] == 0) {
@@ -219,7 +223,7 @@ namespace Dune {
           }
         }
         else if (mask[i.index()] == 1) {
-          for (ColIterator j = _A_[i.index()].begin(); j != _A_[i.index()].end(); ++j)
+          for (ColIterator j = (*_A_)[i.index()].begin(); j != (*_A_)[i.index()].end(); ++j)
             if (mask[j.index()] != 2)
               (*j).usmv(alpha,x[j.index()],y[i.index()]);
         }
@@ -233,7 +237,7 @@ namespace Dune {
     }
 
   private:
-    const matrix_type& _A_;
+    std::shared_ptr<const matrix_type> _A_;
     const communication_type& communication;
     mutable bool buildcomm;
     mutable std::vector<double> mask;
@@ -263,8 +267,10 @@ namespace Dune {
 
   template<class C, class P>
   class NonoverlappingBlockPreconditioner
-    : public Dune::Preconditioner<typename P::domain_type,typename P::range_type> {
+    : public Preconditioner<typename P::domain_type,typename P::range_type> {
     friend struct Amg::ConstructionTraits<NonoverlappingBlockPreconditioner<C,P> >;
+    using X = typename P::domain_type;
+    using Y = typename P::range_type;
   public:
     //! \brief The domain type of the preconditioner.
     typedef typename P::domain_type domain_type;
@@ -280,9 +286,27 @@ namespace Dune {
        \param c The communication object for syncing owner and copy
        data points. (E.~g. OwnerOverlapCommunication )
      */
-    NonoverlappingBlockPreconditioner (P& prec, const communication_type& c)
-      : preconditioner(prec), communication(c)
-    {}
+    /*! \brief Constructor.
+
+       constructor gets all parameters to operate the prec.
+       \param p The sequential preconditioner.
+       \param c The communication object for syncing overlap and copy
+       data points. (E.~g. OwnerOverlapCopyCommunication )
+     */
+    NonoverlappingBlockPreconditioner (P& p, const communication_type& c)
+      : _preconditioner(stackobject_to_shared_ptr(p)), _communication(c)
+    {   }
+
+    /*! \brief Constructor.
+
+       constructor gets all parameters to operate the prec.
+       \param p The sequential preconditioner.
+       \param c The communication object for syncing overlap and copy
+       data points. (E.~g. OwnerOverlapCopyCommunication )
+     */
+    NonoverlappingBlockPreconditioner (const std::shared_ptr<P>& p, const communication_type& c)
+      : _preconditioner(p), _communication(c)
+    {   }
 
     /*!
        \brief Prepare the preconditioner.
@@ -291,7 +315,7 @@ namespace Dune {
      */
     virtual void pre (domain_type& x, range_type& b)
     {
-      preconditioner.pre(x,b);
+      _preconditioner->pre(x,b);
     }
 
     /*!
@@ -304,8 +328,15 @@ namespace Dune {
       // block preconditioner equivalent to WrappedPreconditioner from
       // pdelab/backend/ovlpistsolverbackend.hh,
       // but not to BlockPreconditioner from schwarz.hh
-      preconditioner.apply(v,d);
-      communication.addOwnerCopyToOwnerCopy(v,v);
+      _preconditioner->apply(v,d);
+      _communication.addOwnerCopyToOwnerCopy(v,v);
+    }
+
+    template<bool forward>
+    void apply (X& v, const Y& d)
+    {
+      _preconditioner->template apply<forward>(v,d);
+      _communication.copyOwnerToAll(v,v);
     }
 
     /*!
@@ -315,7 +346,7 @@ namespace Dune {
      */
     virtual void post (domain_type& x)
     {
-      preconditioner.post(x);
+      _preconditioner->post(x);
     }
 
     //! Category of the preconditioner (see SolverCategory::Category)
@@ -326,10 +357,10 @@ namespace Dune {
 
   private:
     //! \brief a sequential preconditioner
-    P& preconditioner;
+    std::shared_ptr<P> _preconditioner;
 
     //! \brief the communication object
-    const communication_type& communication;
+    const communication_type& _communication;
   };
 
   /** @} end documentation */