+++ dune/istl/paamg/aggregates.hh
@@ -15,6 +15,7 @@
 #include <dune/common/sllist.hh>
 #include <dune/common/unused.hh>
 #include <dune/common/ftraits.hh>
+#include <dune/common/scalarmatrixview.hh>
 
 #include <utility>
 #include <set>
@@ -379,16 +380,32 @@ namespace Dune
 
       /**
        * @brief compute the norm of a matrix.
-       * @param m The matrix ro compute the norm of.
+       * @param m The matrix to compute the norm of
        */
       template<class M>
-      typename FieldTraits<typename M::field_type>::real_type operator()(const M& m) const
+      typename FieldTraits<typename M::field_type>::real_type operator()(const M& m,
+                                                                         typename std::enable_if_t<!Dune::IsNumber<M>::value>* sfinae = nullptr) const
       {
         typedef typename M::field_type field_type;
         typedef typename FieldTraits<field_type>::real_type real_type;
         static_assert( std::is_convertible<field_type, real_type >::value,
                   "use of diagonal norm in AMG not implemented for complex field_type");
         return m[N][N];
+        // possible implementation for complex types: return signed_abs(m[N][N]);
+      }
+
+      /**
+       * @brief Compute the norm of a scalar
+       * @param m The scalar to compute the norm of
+       */
+      template<class M>
+      auto operator()(const M& m,
+                      typename std::enable_if_t<Dune::IsNumber<M>::value>* sfinae = nullptr) const
+      {
+        typedef typename FieldTraits<M>::real_type real_type;
+        static_assert( std::is_convertible<M, real_type >::value,
+                  "use of diagonal norm in AMG not implemented for complex field_type");
+        return m;
         // possible implementation for complex types: return signed_abs(m[N][N]);
       }
 
@@ -490,11 +507,8 @@ namespace Dune
     /**
      * @brief Criterion taking advantage of symmetric matrices.
      *
-     * The two template parameters are:
-     * <dl>
-     * <dt>M</dt> <dd>The type of the matrix the amg coarsening works on, e. g. BCRSMatrix</dd>
-     * <dt>Norm</dt> <dd>The norm to use to determine the strong couplings between the nodes, e.g. FirstDiagonal or RowSum.</dd>
-     * </dl>
+     * \tparam M The type of the matrix the amg coarsening works on, e.g. BCRSMatrix
+     * \tparam Norm The norm to use to determine the strong couplings between the nodes, e.g. FirstDiagonal or RowSum.
      */
     template<class M, class Norm>
     class SymmetricCriterion : public AggregationCriterion<SymmetricDependency<M,Norm> >
@@ -509,15 +523,12 @@ namespace Dune
 
 
     /**
-     * @brief Criterion suited for unsymmetric matrices.
+     * @brief Criterion suitable for unsymmetric matrices.
      *
      * Nevertheless the sparsity pattern has to be symmetric.
      *
-     * The two template parameters are:
-     * <dl>
-     * <dt>M</dt> <dd>The type of the matrix the amg coarsening works on, e. g. BCRSMatrix</dd>
-     * <dt>Norm</dt> <dd>The norm to use to determine the strong couplings between the nodes, e.g. FirstDiagonal or RowSum.</dd>
-     * </dl>
+     * \tparam M The type of the matrix the amg coarsening works on, e.g. BCRSMatrix
+     * \tparam Norm The norm to use to determine the strong couplings between the nodes, e.g. FirstDiagonal or RowSum.
      */
     template<class M, class Norm>
     class UnSymmetricCriterion : public AggregationCriterion<Dependency<M,Norm> >
@@ -625,13 +636,11 @@ namespace Dune
       /**
        * @brief Breadth first search within an aggregate
        *
-       * The template parameters: <br>
-       * <dl><dt>reset</dt><dd>If true the visited flags of the vertices
-       *  will be reset after
-       * the search</dd>
-       * <dt>G</dt><dd>The type of the graph we perform the search on.</dd>
-       * <dt>F</dt><dd>The type of the visitor to operate on the vertices</dd>
-       * </dl>
+       * \tparam reset If true the visited flags of the vertices
+       *  will be reset after the search
+       * \tparam G The type of the graph we perform the search on
+       * \tparam F The type of the visitor to operate on the vertices
+       *
        * @param start The vertex where the search should start
        * from. This does not need to belong to the aggregate.
        * @param aggregate The aggregate id.
@@ -651,13 +660,13 @@ namespace Dune
       /**
        * @brief Breadth first search within an aggregate
        *
-       * The template parameters: <br>
-       * <dl><dt>L</dt><dd>A container type providing push_back(Vertex), and
-       * pop_front() in case remove is true</dd>
-       * <dt>remove</dt><dd> If true the entries in the visited list
-       * will be removed.</dd>
-       * <dt>reset</dt><dd>If true the visited flag will be reset after
-       * the search</dd></dl>
+       * \tparam L A container type providing push_back(Vertex), and
+       * pop_front() in case remove is true
+       * \tparam remove If true the entries in the visited list
+       * will be removed.
+       * \tparam reset If true the visited flag will be reset after
+       * the search
+       *
        * @param start The vertex where the search should start
        * from. This does not need to belong to the aggregate.
        * @param aggregate The aggregate id.
@@ -924,8 +933,8 @@ namespace Dune
       /**
        * @brief Build the aggregates.
        *
-       * The template parameter C Is the type of the coarsening Criterion to
-       * use.
+       * \tparam C The type of the coarsening Criterion to use
+       *
        * @param m The matrix to build the aggregates accordingly.
        * @param graph A (sub) graph of the matrix.
        * @param aggregates Aggregate map we will build. All entries should be initialized
@@ -1821,7 +1830,7 @@ namespace Dune
           for(ColIterator col = row.begin(); col != end; ++col)
             if(col.index()!=*vertex) {
               criterion.examine(col);
-              absoffdiag = max(absoffdiag, col->frobenius_norm());
+              absoffdiag = max(absoffdiag, Impl::asMatrix(*col).frobenius_norm());
             }
 
           if(absoffdiag==0)